lambda/chat_handler/pyproject.toml
[project]
name = "chat-handler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "requests>=2.32.5",
    "strands-agents>=1.5.0",
]
